requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# ML
torch==2.3.0
torchvision==0.18.0

# Image processing
opencv-python-headless==4.9.0.80
Pillow==10.3.0
scikit-image==0.23.2
numpy==1.26.4

# Frontend
streamlit==1.35.0
requests==2.32.3
# Logging
loguru==0.7.3